Issue#882:  Moving rule concurrent_007 to conditional_waveforms_001.
