+++ vsg/rules/conditional_waveforms/__init__.py
@@ -1,4 +1,6 @@
 
+from .rule_001 import rule_001
+
 from .rule_100 import rule_100
 from .rule_101 import rule_101
 from .rule_102 import rule_102
